refactor: moved List selection dialog to core module, updated resource references
