+++ app/logbook/olog/ui/src/test/java/org/phoebus/logbook/olog/ui/ListSelectionDemo.java
@@ -4,6 +4,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
+import org.phoebus.ui.dialog.ListSelectionController;
 import org.phoebus.ui.javafx.ApplicationWrapper;
 
 import java.io.IOException;
@@ -20,7 +21,7 @@ public class ListSelectionDemo extends ApplicationWrapper {
     public void start(Stage primaryStage) throws IOException {
         FXMLLoader loader = new FXMLLoader();
 
-        loader.setLocation(this.getClass().getResource("ListSelection.fxml"));
+        loader.setLocation(this.getClass().getResource("/org/phoebus/ui/dialog/ListSelection.fxml"));
         loader.load();
         ListSelectionController controller = loader.getController();
         controller.setAvailable(Arrays.asList("a", "b", "c"));
